+++ packages/expo-file-system/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 🐛 Bug fixes
 
+- [Android] Fix copy/move support for SAF and content provider URIs.
+
 ### 💡 Others
 
 ## 55.0.4 — 2026-02-03